+++ paddleslim/nas/ofa/ofa.py
@@ -20,11 +20,12 @@ import paddle.fluid as fluid
 from .utils.utils import get_paddle_version, remove_model_fn
 pd_ver = get_paddle_version()
 if pd_ver == 185:
-    from .layers_old import BaseBlock, SuperConv2D, SuperLinear
+    from .layers_old import SuperConv2D, SuperLinear
     Layer = paddle.fluid.dygraph.Layer
 else:
-    from .layers import BaseBlock, SuperConv2D, SuperLinear
+    from .layers import SuperConv2D, SuperLinear
     Layer = paddle.nn.Layer
+from .layers_base import BaseBlock
 from .utils.utils import search_idx
 from ...common import get_logger
 from ...core import GraphWrapper, dygraph2program
